move _memo function to misc and implement for subsurface_vertex_inverse
